% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwgroundwaterdata.R
\docType{data}
\name{mwgroundwaterdata}
\alias{mwgroundwaterdata}
\title{mwgroundwaterdata: Malawi Groundwater Monitoring Time-Series Dataset (2024–2025)}
\format{
A tibble with 1415 rows and 9 variables
\describe{
\item{date}{Date when data was captured}
\item{waterpoint_name}{The name of the water point}
\item{district}{Administrative district the water point is located}
\item{latitude}{GPS latitude coordinate}
\item{longitude}{GPS longitude coordinate}
\item{source}{The device that captured the information}
\item{water_level}{Water level of the water}
\item{temperature}{Temperature of the data}
\item{conductivity}{Electical conductivity of the water}
}
}
\usage{
mwgroundwaterdata
}
\description{
This dataset contains groundwater monitoring data collected from monitoring wells across 10 districts in Malawi between 2024 and 2025. The data were captured using automated data loggers installed in monitoring wells and were collected and managed by BASEflow.
The dataset provides time-series measurements of key groundwater parameters, enabling detailed analysis of aquifer behavior across multiple geographic locations.
}
\keyword{datasets}
